Use SciMLTesting v1.2 folder-based run_tests
Convert the test suite to the SciMLTesting v1.2 folder-discovery model:

- runtests.jl is now `using SciMLTesting; run_tests()`. The folder layout is
  the single source of dispatch:
    * Core = top-level test/*.jl (sampler_tests, bigfloat_tests,
      passthrough_rng_tests, alloc_tests), run with the main test env.
    * QA   = test/qa/ (its own Project.toml sub-env, activated by run_tests).
  test_groups.toml is unchanged (Core, QA).
- Extract the previously-inline non-QA testsets into self-contained top-level
  Core files (sampler_tests.jl, bigfloat_tests.jl, passthrough_rng_tests.jl);
  alloc_tests.jl was already a file. Each is run in its own @safetestset by the
  harness, so the exact set of tests per GROUP is preserved.
- Add SciMLTesting + SafeTestsets to the root test deps ([extras]+[targets].test,
  with [compat]) and to test/qa/Project.toml. Drop Pkg from the root test deps
  and its compat (only the old hand-written QA-activation harness used it).
