+++ Project.toml
@@ -14,9 +14,10 @@ Distributions = "0.25"
 ExplicitImports = "1.14.0"
 JET = "0.9, 0.10, 0.11"
 LogExpFunctions = "0.3.14, 1.0"
-Pkg = "1"
 PrecompileTools = "1.2.0"
 Random = "1.10"
+SafeTestsets = "0.1, 1"
+SciMLTesting = "1"
 Statistics = "1"
 Test = "1"
 julia = "1.10"
@@ -27,9 +28,10 @@ Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AllocCheck", "Aqua", "Statistics", "Test", "Distributions", "JET", "ExplicitImports", "Pkg"]
+test = ["AllocCheck", "Aqua", "Statistics", "Test", "Distributions", "JET", "ExplicitImports", "SafeTestsets", "SciMLTesting"]
